// Timetic-developer-frontend/src/app/network/model/commonUser.ts
/**
 * Timetic fejlesztői backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: you@your-company.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CommonUser { 
    id?: string;
    name?: string;
    email?: string;
    registrationDate?: number;
}